+++ Example/Controllers/SwiftUI/Features/Settings/Features/MiniAppSettingsContactsView.swift
@@ -110,10 +110,3 @@ extension MiniAppSettingsContactsView {
         }
     }
 }
-
-func ??<T>(lhs: Binding<Optional<T>>, rhs: T) -> Binding<T> {
-    Binding(
-        get: { lhs.wrappedValue ?? rhs },
-        set: { lhs.wrappedValue = $0 }
-    )
-}
