+++ Example/Controllers/SwiftUI/Features/Settings/MiniAppSettingsView.swift
@@ -135,34 +135,6 @@ struct MiniAppSettingsView: View {
             .font(.system(size: 16, weight: .medium))
 
             Text(viewModel.getBuildVersionText())
-//            Section(header: Text("Single")) {
-//                MiniAppFeatureConfigIdCell(
-//                    store: store,
-//                    placeholder: "MiniApp Id",
-//                    miniAppId: store.$miniAppIdentifierSingle,
-//                    version: store.$miniAppVersionSingle
-//                )
-//            }
-//            Section(header: Text("Tripple")) {
-//                MiniAppFeatureConfigIdCell(
-//                    store: store,
-//                    placeholder: "MiniApp Id (1)",
-//                    miniAppId: store.$miniAppIdentifierTrippleFirst,
-//                    version: store.$miniAppVersionTrippleFirst
-//                )
-//                MiniAppFeatureConfigIdCell(
-//                    store: store,
-//                    placeholder: "MiniApp Id (2)",
-//                    miniAppId: store.$miniAppIdentifierTrippleSecond,
-//                    version: store.$miniAppVersionTrippleSecond
-//                )
-//                MiniAppFeatureConfigIdCell(
-//                    store: store,
-//                    placeholder: "MiniApp Id (3)",
-//                    miniAppId: store.$miniAppIdentifierTrippleThird,
-//                    version: store.$miniAppVersionTrippleThird
-//                )
-//            }
         }
         .font(.system(size: 13))
         .navigationTitle("Settings")
@@ -203,10 +175,6 @@ struct MiniAppSettingsView: View {
                 ()
             }
         }
-        .onAppear {
-            print(config)
-            print("-")
-        }
     }
 }
 
@@ -219,63 +187,5 @@ struct MiniAppAlertMessage: Identifiable {
 struct MiniAppFeatureConfigView_Previews: PreviewProvider {
     static var previews: some View {
         MiniAppSettingsView(store: MiniAppStore.empty(), showFullProgress: .constant(false))
-    }
-}
-
-struct MiniAppSettingsListCellView: View {
-    
-    @State var text: String
-    @State var image: Image
-    
-    var body: some View {
-        Label {
-            Text(text)
-        } icon: {
-            image
-            .font(.system(size: 16, weight: .bold))
-            .foregroundColor(.red)
-        }
-        .padding(.vertical, 15)
-    }
-}
-
-struct MiniAppFeatureConfigIdCell: View {
-
-    @ObservedObject var store: MiniAppStore
-
-    @State var placeholder: String
-    @Binding var miniAppId: String
-    @Binding var version: String
-    @State var isPickerPresented: Bool = false
-
-    var body: some View {
-        HStack(spacing: 10) {
-            VStack {
-                TextField(placeholder, text: $miniAppId)
-                    .frame(height: 30)
-                    .frame(maxWidth: .infinity)
-                TextField("Version", text: $version)
-                    .frame(height: 30)
-                    .frame(maxWidth: .infinity)
-            }
-
-            Button {
-                isPickerPresented = true
-            } label: {
-                Image(systemName: "list.dash")
-            }
-            .frame(width: 50, height: 50)
-            .background(RoundedRectangle(cornerRadius: 10).fill(Color(.secondarySystemBackground)))
-            .sheet(isPresented: $isPickerPresented, content: {
-                MiniAppListPickerView(
-                    store: store,
-                    isPresented: $isPickerPresented,
-                    miniAppId: $miniAppId,
-                    version: $version
-                )
-                .environmentObject(store)
-            })
-        }
-        .padding(.vertical, 10)
     }
 }
