+++ Example/Controllers/SwiftUI/Helper/Binding+Optional.swift
@@ -1,0 +1,9 @@
+import Foundation
+import SwiftUI
+
+func ??<T>(lhs: Binding<Optional<T>>, rhs: T) -> Binding<T> {
+    Binding(
+        get: { lhs.wrappedValue ?? rhs },
+        set: { lhs.wrappedValue = $0 }
+    )
+}
