+++ Example/Controllers/SwiftUI/MiniAppDashboardView.swift
@@ -7,9 +7,7 @@ struct MiniAppDashboardView: View {
 
     @State var sampleMiniAppId: String = ""
     @State var sampleMiniAppVersion: String = ""
-
     @State var selection: Int = 0
-    
     @State var isPresentingFullProgress: Bool = false
 
     var body: some View {
@@ -43,16 +41,6 @@ struct MiniAppDashboardView: View {
                 }
                 .tag(2)
 
-    //            MiniAppSingleView(
-    //                miniAppId: sampleMiniAppId,
-    //                miniAppVersion: sampleMiniAppVersion,
-    //                miniAppType: .miniapp
-    //            )
-    //            .tabItem {
-    //                Label("Sample", systemImage: "cube")
-    //            }
-    //            .tag(3)
-
                 NavigationView {
                     MiniAppSettingsView(store: store, showFullProgress: $isPresentingFullProgress)
                 }
@@ -60,7 +48,7 @@ struct MiniAppDashboardView: View {
                 .tabItem {
                     Label("Settings", systemImage: "gear")
                 }
-                .tag(4)
+                .tag(3)
             }
 
             if isPresentingFullProgress {
@@ -74,7 +62,7 @@ struct MiniAppDashboardView: View {
         .navigationTitle(navigationTitle)
         .onAppear {
             if !store.miniAppSetupCompleted {
-                selection = 4
+                selection = 3
             }
         }
     }
@@ -88,8 +76,6 @@ struct MiniAppDashboardView: View {
         case 2:
             return "MiniApp Features"
         case 3:
-            return "MiniApp Sample"
-        case 4:
             return "Settings"
         default:
             return "Unknown"
